move: ユーザー管理ページのコンポーネントのファイルをcomponentsからcomponents/user-componentsに移動しました。
