+++ components/index.ts
@@ -2,8 +2,3 @@ export { MenuItemIcon } from "./MenuItemIcon";
 export { MenuCard } from "./MenuCard";
 export { AddMenuCard } from "./AddMenuCard";
 export { OrderCard } from "./OrderCard";
-export { AddUserCard } from "./AddUserCard";
-export { UserCard } from "./UserCard";
-export { UserCardOutline } from "./UserCardOutline";
-export { UserCardTitleSpace } from "./UserCardTitleSpace";
-export { UserCardTitle } from "./UserCardTitle";
