+++ components/user-components/UserCardOutline.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
 import classNames from "classnames";
 
-import { UserCardItemProps } from "../types";
+import { UserCardItemProps } from "../../types";
 
 interface Props extends UserCardItemProps {
   children: ReactNode;
